+++ source/Tubeshade.Data/Migrations/000047_view-status.sql
@@ -1,0 +1,1 @@
+﻿CREATE TYPE media.view_status AS ENUM ('viewed', 'not_viewed', 'partially_viewed');