+++ source/Tubeshade.Server/Pages/Index.cshtml.cs
@@ -48,7 +48,7 @@ public sealed class IndexModel : PageModel, IVideoPage, INonLibraryPage
 
     /// <inheritdoc />
     [BindProperty(SupportsGet = true)]
-    public bool? Viewed { get; set; }
+    public ViewStatus? Viewed { get; set; }
 
     /// <inheritdoc />
     [BindProperty(SupportsGet = true)]