+++ source/Tubeshade.Server/Pages/Libraries/Library.cshtml.cs
@@ -50,7 +50,7 @@ public sealed class Library : LibraryPageBase, IVideoPage, IPageWithSettings
 
     /// <inheritdoc />
     [BindProperty(SupportsGet = true)]
-    public bool? Viewed { get; set; }
+    public ViewStatus? Viewed { get; set; }
 
     /// <inheritdoc />
     [BindProperty(SupportsGet = true)]